pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coyin"
version = "0.1.0"
description = "A research-oriented desktop workspace built with Qt."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.7",
  "requests>=2.31",
  "pypdf>=4.2",
  "python-docx>=1.1",
  "markdown-it-py>=3.0",
  "PyYAML>=6.0"
]

[project.scripts]
coyin = "coyin.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
